feat: move excel parsing to separate class
